backend works for adding city

+++ AP.BTP.Application/CQRS/CityDTO.cs
@@ -11,7 +11,7 @@ namespace AP.BTP.Application.CQRS
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Population { get; set; }
+        public long Population { get; set; }
 
         public int CountryId { get; set; }
     }
